+++ packages/uql-orm/src/mysql/mysql2QuerierPool.test.ts
@@ -1,9 +1,9 @@
-import { AbstractQuerierPoolIt } from '../querier/abstractQuerierPool-test.js';
+import { AbstractSqlQuerierPoolIt } from '../querier/abstractSqlQuerierPool-test.js';
 import { createSpec } from '../test/index.js';
 import type { MySql2Querier } from './mysql2Querier.js';
 import { MySql2QuerierPool } from './mysql2QuerierPool.js';
 
-export class MySql2QuerierPoolIt extends AbstractQuerierPoolIt<MySql2Querier> {
+export class MySql2QuerierPoolIt extends AbstractSqlQuerierPoolIt<MySql2Querier> {
   constructor() {
     super(
       new MySql2QuerierPool({