+++ packages/uql-orm/src/mysql/mysql2QuerierPool.ts
@@ -1,10 +1,10 @@
 import { createPool, type Pool, type PoolOptions } from 'mysql2/promise';
-import { AbstractQuerierPool } from '../querier/index.js';
+import { AbstractSqlQuerierPool } from '../querier/index.js';
 import type { ExtraOptions } from '../type/index.js';
 import { MySql2Dialect } from './mysql2Dialect.js';
 import { MySql2Querier } from './mysql2Querier.js';
 
-export class MySql2QuerierPool extends AbstractQuerierPool<MySql2Dialect, MySql2Querier> {
+export class MySql2QuerierPool extends AbstractSqlQuerierPool<MySql2Querier, MySql2Dialect> {
   readonly pool: Pool;
 
   constructor(opts: PoolOptions, extra?: ExtraOptions) {