+++ packages/uql-orm/src/sqlite/sqliteQuerierPool.test.ts
@@ -1,9 +1,9 @@
-import { AbstractQuerierPoolIt } from '../querier/abstractQuerierPool-test.js';
+import { AbstractSqlQuerierPoolIt } from '../querier/abstractSqlQuerierPool-test.js';
 import { createSpec } from '../test/index.js';
 import type { SqliteQuerier } from './sqliteQuerier.js';
 import { Sqlite3QuerierPool } from './sqliteQuerierPool.js';
 
-export class Sqlite3QuerierPoolIt extends AbstractQuerierPoolIt<SqliteQuerier> {
+export class Sqlite3QuerierPoolIt extends AbstractSqlQuerierPoolIt<SqliteQuerier> {
   constructor() {
     super(new Sqlite3QuerierPool(':memory:'));
   }